extract navlinkspanel to own component
